# openapi.yaml
openapi: 3.0.3
info:
  title: full-stack-framework API
  description: Automatically generated OpenAPI spec from Postman collection "full-stack-framework"
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Default local server
paths:
...
  /api/test/testcase:
    get:
      summary: Get testcases
      operationId: getTestcases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTestcasesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/test/tasks:
    get:
      summary: Get tasks
      operationId: getTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetTasks'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/test/task:
    get:
      summary: Get task
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Task ID
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
            example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetTask'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

    post:
      summary: Submit task
      operationId: submitTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestSubmitTask'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

    delete:
      summary: Cancel task
      operationId: cancelTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Task ID
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
            example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/test/testlog:
    get:
      summary: Get test log
      operationId: getTestLog
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Task ID
          required: true
          schema:
            type: integer
            format: uint64
            minimum: 1
            example: 74
        - name: testName
          in: query
          description: Test name
          required: true
          schema:
            type: string
            example: TestRegistration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetTestLog'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/admin/test/testcase:
    post:
      summary: Add testcases
      operationId: addTestcases
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTestcasesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

    delete:
      summary: Delete testcases
      operationId: deleteTestcases
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTestcasesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /api/admin/test/history:
    delete:
      summary: Delete tasks history
      operationId: deleteTasksHistory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    Testcase:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: test-1
        link:
          type: string
          example: https://github.com/free5gc/free5gc#test-1
    GetTestcasesResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Testcases retrieved successfully
        testcases:
          type: array
          items:
            $ref: '#/components/schemas/Testcase'
    TaskSimple:
      type: object
      required:
        - id
        - username
        - createTime
      properties:
        id:
          type: integer
          format: uint64
          example: 1
        username:
          type: string
          example: admin
        createTime:
          type: integer
          format: int64
          example: 1777143861
    NfPr:
      type: object
      required:
        - nfName
        - pr
      properties:
        nfName:
          type: string
          example: amf
        pr:
          type: integer
          example: 100
    LibraryPr:
      type: object
      required:
        - repoName
        - pr
      properties:
        repoName:
          type: string
          enum:
            - openapi
            - util
            - nas
            - ngap
            - pfcp
            - aper
          example: openapi
        pr:
          type: integer
          example: 67
...
    ResponseGetTasks:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Tasks retrieved successfully
        pendingTask:
          type: array
          items:
            $ref: '#/components/schemas/TaskSimple'
        ongoingTask:
          type: array
          items:
            $ref: '#/components/schemas/TaskSimple'
        historyTask:
          type: array
          items:
            $ref: '#/components/schemas/TaskSimple'
    ResponseGetTask:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Task retrieved successfully
        id:
          type: integer
          format: uint64
          example: 1
        username:
          type: string
          example: admin
        createTime:
          type: integer
          format: int64
          example: 1777143861
        status:
          type: string
          enum:
            - pending
            - running
            - success
            - failed
            - timeout
            - canceled
          example: success
        tests:
          type: array
          items:
            $ref: '#/components/schemas/TestDetail'
        nfPrList:
          type: array
          items:
            $ref: '#/components/schemas/NfPr'
        libraryPrList:
          type: array
          items:
            $ref: '#/components/schemas/LibraryPr'
    TestDetail:
      type: object
      required:
        - name
        - status
      properties:
        name:
          type: string
          example: TestRegistration
        status:
          type: string
          example: pending
    ResponseGetTestLog:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Test log retrieved successfully
        log:
          type: string
          example: test log content
    RequestSubmitTask:
      type: object
      required:
        - tests
        - nfPrList
      properties:
        tests:
          type: array
          items:
            type: string
          example:
            - test-1
            - test-2
        nfPrList:
          type: array
          items:
            $ref: '#/components/schemas/NfPr'
        libraryPrList:
          type: array
          items:
            $ref: '#/components/schemas/LibraryPr'
    AddTestcasesRequest:
      type: object
      required:
        - testcases
      properties:
        testcases:
          type: array
          items:
            $ref: '#/components/schemas/Testcase'
    DeleteTestcasesRequest:
      type: object
      required:
        - testcases
      properties:
        testcases:
          type: array
          items:
            $ref: '#/components/schemas/Testcase'
...